+++ src/components/spartan/SSidebar/index.ts
@@ -1,0 +1,1 @@
+export { default as SSidebar } from './SSidebar.vue';