+++ src/components/spartan/SSidebar/types.ts
@@ -1,0 +1,13 @@
+import type { FunctionalComponent } from 'vue';
+
+export type TSidebarProps = {
+    placetopayHeader: boolean;
+};
+
+export type TSidebarItemProps = {
+    path: string;
+    icon: FunctionalComponent;
+};
+export type TStateDefinition = {
+    path: string;
+};